update App.jsx to save profile in state to then pass it on as props for the respective routes
