+++ src/components/NavBar/NavBar.jsx
@@ -10,7 +10,7 @@ const NavBar = ({ user, handleLogout }) => {
           <li><NavLink to="/profiles">Profiles</NavLink></li>
           <li><NavLink to="" onClick={handleLogout}>LOG OUT</NavLink></li>
           <li><NavLink to="/auth/change-password">Change Password</NavLink></li>
-          <li><NavLink to={`/profiles/${user.profile}`}>My Profile</NavLink></li>
+          <li><NavLink to="/my-profile">My Profile</NavLink></li>
         </ul>
       :
         <ul>
