+++ src/pages/AddFavorites/AddFavorites.jsx
@@ -1,6 +1,5 @@
 // npm modules
-import { useState, useEffect } from "react"
-import { useParams } from "react-router-dom"
+import { useState } from "react"
 
 //components
 import SearchForm from "../../components/SearchForm/SearchForm"
@@ -11,7 +10,6 @@ import * as tmdbService from '../../services/tmdbService'
 
 const AddFavorites = () => {
   const [celebs, setCelebs] = useState(null)
-  const {profileId} = useParams()
 
   const searchCelebs = async (formData) => {
       const celebsData = await tmdbService.getCelebs(formData)
